lab1/requirements.txt
numpy==1.16.1
PyOpenGL==3.1.0
